Moved functions to helper
